+++ src/app/containers/EnterPage/enter.scss
